+++ alembic/versions/31c4968699ad_add_etl_content_to_records.py
@@ -1,0 +1,55 @@
+"""add etl content to records
+
+Revision ID: 31c4968699ad
+Revises: 9d5fb67e29f7
+Create Date: 2025-11-23 23:08:27.327070
+
+"""
+
+from alembic import op
+import sqlalchemy as sa
+
+
+# revision identifiers, used by Alembic.
+revision = "31c4968699ad"
+down_revision = "9d5fb67e29f7"
+branch_labels = None
+depends_on = None
+
+
+def upgrade():
+    # ### commands auto generated by Alembic - please adjust! ###
+    op.add_column(
+        "etl_task",
+        sa.Column("original_file_name", sa.String(), nullable=True),
+        schema="global",
+    )
+    op.add_column(
+        "github_file",
+        sa.Column("content", sa.String(), nullable=True),
+        schema="integration",
+    )
+    op.add_column(
+        "github_issue",
+        sa.Column("content", sa.String(), nullable=True),
+        schema="integration",
+    )
+    op.add_column(
+        "pdf", sa.Column("content", sa.String(), nullable=True), schema="integration"
+    )
+    op.add_column(
+        "sharepoint",
+        sa.Column("content", sa.String(), nullable=True),
+        schema="integration",
+    )
+    # ### end Alembic commands ###
+
+
+def downgrade():
+    # ### commands auto generated by Alembic - please adjust! ###
+    op.drop_column("sharepoint", "content", schema="integration")
+    op.drop_column("pdf", "content", schema="integration")
+    op.drop_column("github_issue", "content", schema="integration")
+    op.drop_column("github_file", "content", schema="integration")
+    op.drop_column("etl_task", "original_file_name", schema="global")
+    # ### end Alembic commands ###
