--- alembic/versions/ab353410347b_add_etl_original_file_name.py
@@ -1,33 +1,0 @@
-"""add etl original file name
-
-Revision ID: ab353410347b
-Revises: 9d5fb67e29f7
-Create Date: 2025-11-22 21:43:41.539486
-
-"""
-
-from alembic import op
-import sqlalchemy as sa
-
-
-# revision identifiers, used by Alembic.
-revision = "ab353410347b"
-down_revision = "9d5fb67e29f7"
-branch_labels = None
-depends_on = None
-
-
-def upgrade():
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.add_column(
-        "etl_task",
-        sa.Column("original_file_name", sa.String(), nullable=True),
-        schema="global",
-    )
-    # ### end Alembic commands ###
-
-
-def downgrade():
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.drop_column("etl_task", "original_file_name", schema="global")
-    # ### end Alembic commands ###
